ENH: Move UndefinedVariableError to error/__init__.py per GH27656
