+++ pandas/core/computation/scope.py
@@ -15,6 +15,7 @@ import numpy as np
 
 from pandas._libs.tslibs import Timestamp
 from pandas.compat.chainmap import DeepChainMap
+from pandas.errors import UndefinedVariableError
 
 
 def ensure_scope(
@@ -207,9 +208,6 @@ class Scope:
                 # e.g., df[df > 0]
                 return self.temps[key]
             except KeyError as err:
-                # runtime import because ops imports from scope
-                from pandas.core.computation.ops import UndefinedVariableError
-
                 raise UndefinedVariableError(key, is_local) from err
 
     def swapkey(self, old_key: str, new_key: str, new_value=None) -> None:
